tests/wpt/tests/css/css-overflow/scroll-markers/scroll-marker-group-size-container-query.html

<!DOCTYPE html>
<title>CSS Overflow Test: Size container for ::scroll-marker-group</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#scroll-marker-group-property">
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 400px;
  }
  #scroller {
    container-type: inline-size;
    overflow: auto;
    width: 200px;
    height: 200px;
    scroll-marker-group: before;
    @container (width = 400px) {
      &::scroll-marker-group { --test: pass; }
    }
  }
</style>
<div id="container">
  <div id="scroller"></div>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(scroller, "::scroll-marker-group").getPropertyValue("--test"), "pass");
  }, "::scroll-marker-group skips originating element for size queries");
</script>
